Define plugin and refactoring.  - extract specialCharacter module to plugin.  - plugin is a kind of external module.  - extract getSelectedText method from video and special character.
